# Conflicts:
#	app/build.gradle
#	app/src/development/java/com/namoadigital/prj001/ui/AppBase.java
#	app/src/main/java/com/namoadigital/prj001/model/GeOsDeviceItem.kt
#	app/src/main/java/com/namoadigital/prj001/model/GeOsDeviceMaterial.kt
#	app/src/main/java/com/namoadigital/prj001/model/MD_Product_Serial_Tp_Device.kt
#	app/src/main/java/com/namoadigital/prj001/model/MD_Product_Serial_Tp_Device_Item.kt
#	app/src/main/java/com/namoadigital/prj001/model/MD_Product_Serial_Tp_Device_Item_Hist.kt
#	app/src/main/java/com/namoadigital/prj001/model/SO_Pack_Express.java
#	app/src/main/java/com/namoadigital/prj001/model/SO_Pack_Express_Local.java
#	app/src/main/java/com/namoadigital/prj001/model/TK_Ticket.java
#	app/src/main/java/com/namoadigital/prj001/model/TSO_Next_Orders_Env.java
#	app/src/main/java/com/namoadigital/prj001/model/TSO_Pack_Express_Env.java
#	app/src/main/java/com/namoadigital/prj001/model/TSO_Service_Search_Env.java
#	app/src/main/java/com/namoadigital/prj001/model/TSO_Service_Search_Obj.java
#	app/src/main/java/com/namoadigital/prj001/model/TkTicketCache.kt
#	app/src/main/java/com/namoadigital/prj001/ui/act070/Act070_Main.java
#	app/src/main/java/com/namoadigital/prj001/util/ToolBox_Inf.java

+++ app/src/main/java/com/namoadigital/prj001/service/WS_TK_Ticket_Save.java
@@ -262,6 +262,11 @@ public class WS_TK_Ticket_Save extends IntentService {
                 removeCancelledProducts(tk_ticket);
             }
             //
+            if (tk_ticket.getUpdate_required_status() == 1) {
+                tk_ticket.setTime_action(ConstantBaseApp.ACTION_NOT_EXECUTED);
+                tk_ticket.setChange_date(tk_ticket.getNot_executed_date());
+            }
+            //
             if (tk_ticket.getStep() != null && tk_ticket.getStep().size() > 0) {
                 ArrayList<TK_Ticket_Step> stepsToUpdate = new ArrayList<>();
                 for (TK_Ticket_Step tk_ticket_step : tk_ticket.getStep()) {
@@ -329,6 +334,8 @@ public class WS_TK_Ticket_Save extends IntentService {
         );
         //
         ToolBox.sendBCStatus(getApplicationContext(), "STATUS", hmAux_Trans.get("generic_receiving_data_msg"), "", "0");
+        //
+        ToolBox_Con.setBooleanPreference(getApplicationContext(), ConstantBaseApp.PREFERENCE_SERIAL_OFFLINE_FLOW, false);
         //
         T_TK_Ticket_Save_Rec rec = gsonRec.fromJson(
                 resultado,
@@ -553,6 +560,7 @@ public class WS_TK_Ticket_Save extends IntentService {
                                 ticket.setTicket_status(ConstantBase.SYS_STATUS_CANCELLED);
                                 ticket.setUpdate_required(0);
                                 ticket.setUpdate_required_product(0);
+                                ticket.setUpdate_required_status(0);
 
                                 ticketDao.addUpdate(ticket);
 
@@ -780,35 +788,38 @@ public class WS_TK_Ticket_Save extends IntentService {
         //Executa loop nos tickets caso exista
         if (rec.getTicket() != null && rec.getTicket().size() > 0) {
             for (TK_Ticket tk_ticket : rec.getTicket()) {
-                //Seta pk nos filhos
-                tk_ticket.setPK();
-                //Verifica se precisa resetar alguma foto. Isso deve ser feito se o "file_code" da foto
-                //for alterado, o que significa que mudaram a foto no server...
-                TK_Ticket.checkActionPhotoResetNeeds(
-                        getDbTicket(tk_ticket, false),
-                        tk_ticket
-                );
-                //Varre todas as imagens verificando se existe imagem local para cada item que pode ter foto
-                tk_ticket.updateLocalImagesPathIfExists();
-                /**
-                 * LUCHE - 04/09/2020
-                 * No download do ticket, existe chamado do metodo abaixo, pois como ao criar o form
-                 * os dados de "inicio" do ctrl não são transmitidos e não setam update_required, então
-                 * o download apaga essas infos.
-                 * Aqui, como é save, em tese, não precisa chamada.
-                 * //Busca ctrls tipo form em andamento e que seriam resetados.
-                 * LUCHE - 10/09/2020
-                 * O luche que passado errou feio, precisa sim
-                 */
-                tk_ticket.updateTicketCtrlFormInProcess(getApplicationContext());
-                //Remove o ticket do banco de dados
-                ticketDao.removeFullV2(tk_ticket);
-                //Tenta o insert do ticket
-                DaoObjReturn daoObjReturn = ticketDao.addUpdate(tk_ticket);
-                //Se não houve erro , chama metodo define proximo passo.
+                if(tk_ticket.getTicket_prefix() > 0
+                && tk_ticket.getTicket_code() > 0) {
+                    //Seta pk nos filhos
+                    tk_ticket.setPK();
+                    //Verifica se precisa resetar alguma foto. Isso deve ser feito se o "file_code" da foto
+                    //for alterado, o que significa que mudaram a foto no server...
+                    TK_Ticket.checkActionPhotoResetNeeds(
+                            getDbTicket(tk_ticket, false),
+                            tk_ticket
+                    );
+                    //Varre todas as imagens verificando se existe imagem local para cada item que pode ter foto
+                    tk_ticket.updateLocalImagesPathIfExists();
+                    /**
+                     * LUCHE - 04/09/2020
+                     * No download do ticket, existe chamado do metodo abaixo, pois como ao criar o form
+                     * os dados de "inicio" do ctrl não são transmitidos e não setam update_required, então
+                     * o download apaga essas infos.
+                     * Aqui, como é save, em tese, não precisa chamada.
+                     * //Busca ctrls tipo form em andamento e que seriam resetados.
+                     * LUCHE - 10/09/2020
+                     * O luche que passado errou feio, precisa sim
+                     */
+                    tk_ticket.updateTicketCtrlFormInProcess(getApplicationContext());
+                    //Remove o ticket do banco de dados
+                    ticketDao.removeFullV2(tk_ticket);
+                    //Tenta o insert do ticket
+                    DaoObjReturn daoObjReturn = ticketDao.addUpdate(tk_ticket);
+                    //Se não houve erro , chama metodo define proximo passo.
 //                if (!daoObjReturn.hasError()) {
 //
 //                }
+                }
                 if(!tk_ticket.getSerial().isEmpty()){
                     for(MD_Product_Serial serial: tk_ticket.getSerial()){
                         serialDao.addUpdateTmp(serial);
